Pontoon: Update Russian (ru) localization of AMO

+++ lib/screens/prediction_widget.dart
@@ -17,4 +17,6 @@ class PredictionWidget extends StatelessWidget {
             color: prediction == null
                 ? Colors.black
                 : Colors.red.withOpacity(
-                    (prediction.confidence * 2).clamp(0, 1)
+                    (prediction.confidence * 2).clamp(0, 1).toDouble(),
+                  ),
+
